Start Chat Messaging get room id then go to ActivityChat and send message and finally get response for sended messge

+++ app/src/main/java/com/iGap/request/RequestChatGetRoom.java
@@ -4,7 +4,7 @@ import com.iGap.proto.ProtoChatGetRoom;
 
 public class RequestChatGetRoom {
 
-    public void chatGetRoom(int peerId) {
+    public void chatGetRoom(long peerId) {
 
         ProtoChatGetRoom.ChatGetRoom.Builder chatGetRoom = ProtoChatGetRoom.ChatGetRoom.newBuilder();
         chatGetRoom.setPeerId(peerId);
